Store setup values with the observation
In order to render a historical observation correctly, the setup values need to be persisted with the observation as the setup values will change and are only session values.

Added more fields to the observation including a name field.  Currently I have this working with a default but can be made into an editable field

Cleaned up the views and make sure that creating an observation and editing one is working correctly (minus the videos, they are a separate issue)

Fixed bug so that tags will now save on an observation

Added View Last Observation button to make navigating back easy

Added validation to require at least one construct or student for an observation to be saved

Fixed student and construct choices persisting through errors